refactor: split notify component
